requirements.txt
PyYAML>=6.0
uiautomator2>=3.2.8
mysql-connector-python>=8.0
fastapi>=0.115.0
uvicorn>=0.30.0
